requirements.txt
cffi==2.0.0
contourpy==1.3.3
coverage==7.11.3
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
geographiclib==2.1
geopy==2.4.1
h3==4.3.1
iniconfig==2.3.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pip==24.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
pyswisseph==2.10.3.2
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.2.0
six==1.17.0
timezonefinder==8.1.0
